Major codebase reorganization and modularization
This commit performs extensive restructuring to improve code organization:

**Module reorganization:**
- Split large Widgets.hs into individual widget modules (Alert, Button, Card, Code, Form, Layout, Status)
- Rename Routes.hs to Page/IndexPage.hs for better naming consistency
- Rename Supervisor.hs to Supervisor/Task.hs and extract Core functionality
- Remove Toplevel.hs (functionality distributed to appropriate modules)

**Server architecture refactoring:**
- Create Vira.App.Server module with runServer function containing WAI/Warp/TLS setup
- Extract startWarpServer function to Vira.Lib.TLS for better separation of concerns
- Refactor runServer using where clauses for buildApplication, buildUrl, and staticMiddleware
- Update Run.hs to delegate server startup to the new Server module

**Dependencies and cleanup:**
- Add wai dependency to cabal file
- Update imports across affected modules
- Remove unused language pragmas
- Update CLAUDE.md and DESIGN.md documentation

This reorganization significantly improves modularity, reduces file sizes,
and establishes clearer separation of concerns throughout the codebase.
