+++ CLAUDE.md
@@ -30,7 +30,7 @@ See README.md for general project information.
 
 ## Testing changes
 
-- When possible, just rely on `cabal build` (or HLS). Running the application (via `cabal run` or `nix run`) should reserved for cases where it is necessary. I prefer a more rapid dev-feedback cycle. Remember, with Haskell when code compiles it usually runs correctly as well.
+- When possible, just rely on `cabal build` (or HLS). Running the application (via `cabal run` or `nix run`) should reserved for cases where it is necessary. I prefer a more rapid dev-feedback cycle. Remember, with Haskell when code compiles it usually runs correctly as well. You must fix all GHC warnings introduced by you.
 
 ## Your report
