+++ src/Vira/Page/RegistryPage.hs
@@ -22,7 +22,11 @@ import Vira.Page.RepoPage qualified as RepoPage
 import Vira.State.Acid qualified as St
 import Vira.State.Core qualified as St
 import Vira.State.Type (Repo (..), RepoName (..), RepoSettings (..))
-import Vira.Widgets qualified as W
+import Vira.Widgets.Alert qualified as W
+import Vira.Widgets.Button qualified as W
+import Vira.Widgets.Card qualified as W
+import Vira.Widgets.Form qualified as W
+import Vira.Widgets.Layout qualified as W
 import Prelude hiding (ask, asks, for_)
 
 type FormReq a = ReqBody '[FormUrlEncoded] a
